refactor: 重命名types文件
