+++ App/src/plugins/officialPlugins.ts
@@ -4,7 +4,7 @@
 import timeDisplay from "./timeDisplay/timeDisplay";
 import {scheduleEditor} from "./scheduleEditor/scheduleEditor";
 import {scheduleDisplay} from "./scheduleDisplay/scheduleDisplay";
-import {IPlugin} from "../types/plugins";
+import {IPlugin} from "../types/plugins.types";
 
 const plugins: IPlugin<any>[] = [
     scheduleEditor,
