+++ App/src/plugins/scheduleEditor/components/profileConfig/timetableEditor/timetableLayoutList.vue
@@ -1,6 +1,6 @@
 <script setup lang="ts">
 import {computed} from "vue";
-import {TimetableObject} from "../../../scheduleEditorTypes";
+import {TimetableObject} from "../../../scheduleEditor.types";
 import {v4 as uuidv4} from "uuid";
 import {DateTime, Duration} from "luxon";
 import TcButton from "../../../../../UI/TcButton.vue";
