+++ App/src/types/rsShared.types.ts
